requirements.txt
django
djangorestframework
markdown
arrow
django-bootstrap4
django-markup[all_filter_dependencies]
pylint